Test management yff

+++ src/views/manage-test/manage-exam/index.vue
@@ -234,27 +234,43 @@ export default {
       this.wordsDialog.changeMode = 'add'
     },
     onDeleteClicked(exam_id, exam_index) {
-      deletePaperById({ testOptionId: exam_id }).then(response => {
-        // if (response.data.status === 200) {
-          console.log('delete paper success')
-          this.list.splice(exam_index, 1)
-        // } else {
-        //   alert('删除失败')
-        // }
+      this.$confirm('确认删除本场考试？此操作不可回退！', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deletePaperById({testOptionId: exam_id}).then(response => {
+          if (response.data.status === 200) {
+            // console.log('delete paper success')
+            this.$message({
+              type: 'success',
+              message: '删除成功！'
+            })
+            this.list.splice(exam_index, 1)
+          } else {
+            alert('删除失败')
+          }
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     },
     onEditClicked(exam_id, exam_index) {
       this.examId = exam_id
       this.wordsDialog.title = '编辑考试'
-      this.timeValue = dateFormat(this.list[exam_index].startDate)
+      this.timeValue = this.list[exam_index].startDate
       // console.log("接收时间")
       // console.log(this.timeValue)
       // console.log(typeof this.timeValue)
       let sum = this.list[exam_index].selectNum + this.list[exam_index].judgeNum + this.list[exam_index].qaNum
-      // console.log(sum)
+      this.examType = sum > 0 ? '随机生成' : '勾选生成'
+      console.log(sum)
       // console.log(this.list[paper_index].paperSelectNum)
       if (sum > 0) {
-        this.examType = '随机生成'
+        // this.examType = '随机生成'
         this.form.examSelectNum = this.list[exam_index].selectNum
         this.form.examJudgeNum = this.list[exam_index].judgeNum
         this.form.examQaNum = this.list[exam_index].qaNum
@@ -264,7 +280,7 @@ export default {
         // console.log(this.list[paper_index])
         this.form.duration = this.list[exam_index].duration
       } else {
-        this.examType = '勾选生成'
+        // this.examType = '勾选生成'
         this.index = exam_index
         this.examName = this.list[exam_index].testOptionName
         getPaperIdByTestOptionId(exam_id).then(response => {
@@ -301,7 +317,7 @@ export default {
             judgeNum: this.form.examJudgeNum,
             qaNum: this.form.examQaNum,
             duration: this.form.duration,
-            startDate: dateFormat(this.timeValue)
+            startDate: this.timeValue
           }
           // console.log("写入时间")
           // console.log(this.timeValue)
@@ -314,7 +330,7 @@ export default {
               // judgeNum: this.form.examJudgeNum,
               // qaNum: this.form.examQaNum,
               duration: this.form.duration,
-              startDate: dateFormat(this.timeValue)
+              startDate: this.timeValue
             })
           })
           // this.wordsDialog.visible = false
@@ -331,18 +347,19 @@ export default {
             qaNum: 0,
             duration: this.duration,
             paperId: this.value,
-            startDate: dateFormat(this.timeValue)
+            startDate: this.timeValue
           }
           addPaper(params).then(response => {
             this.list.push({
               testOptionId: response.data.responseMap.result,
               testOptionName: this.examName,
               duration: this.duration,
-              startDate: dateFormat(this.timeValue)
+              startDate: this.timeValue
             })
           })
           // this.wordsDialog.visible = false
         }
+        location.reload();
       } else if (this.wordsDialog.changeMode === 'update') {
         if (this.examType === '随机生成') {
           const exam_data = {
@@ -351,7 +368,7 @@ export default {
             judgeNum: this.form.examJudgeNum,
             qaNum: this.form.examQaNum,
             duration: this.form.duration,
-            startDate: dateFormat(this.timeValue)
+            startDate: this.timeValue
           }
           changePaperById(this.examId, exam_data).then(response => {
             this.list[this.form.index].testOptionName = exam_data.testOptionName
@@ -367,7 +384,7 @@ export default {
             qaNum: 0,
             paperId: this.value,
             duration: this.duration,
-            startDate: dateFormat(this.timeValue)
+            startDate: this.timeValue
           }
           changePaperById(this.examId, exam_data).then(response => {
             this.list[this.index].testOptionName = exam_data.testOptionName
